+++ tests/test_parser_light_and_voc.py
@@ -1,6 +1,6 @@
 import pytest
 
-from pyfreshintellivent.skyModeParser import SkyModeParser
+from pyfreshintellivent import SkyModeParser
 
 parser = SkyModeParser()
